// mobile/src/app/core/models/user.ts
// tslint:disable: variable-name
export class User {
  address: string;
  admin: 0;
  created_at: string;
  deleted_at: string;
  email: string;
  email_verified_at: string;
  firstname: string;
  id: number;
  lastname: string;
  password_reset_at: string;
  password_reset_token: string;
  phone_number: string;
  status_id: number;
  updated_at: string;
  vat_number: number;
}